Move read_config_from_toml() out of lib, add new_with_config() method
