+++ src/cli.rs
@@ -5,7 +5,7 @@ use clap::{Args, Parser, Subcommand};
 use enum_dispatch::enum_dispatch;
 use log::{debug, info};
 use popgetter::{
-    data_request_spec::{BBox, DataRequestSpec, GeometrySpec, MetricSpec, RegionSpec}, formatters::{CSVFormatter, GeoJSONFormatter, GeoJSONSeqFormatter, OutputFormatter, OutputGenerator}, metadata::MetricId, Popgetter
+    config::Config, data_request_spec::{BBox, DataRequestSpec, GeometrySpec, MetricSpec, RegionSpec}, formatters::{CSVFormatter, GeoJSONFormatter, GeoJSONSeqFormatter, OutputFormatter, OutputGenerator}, metadata::MetricId, Popgetter
 };
 use serde::{Deserialize, Serialize};
 use std::fs::File;
@@ -25,7 +25,7 @@ pub enum OutputFormat {
 /// Trait that defines what to run when a given subcommand is invoked.
 #[enum_dispatch]
 pub trait RunCommand {
-    async fn run(&self) -> Result<()>;
+    async fn run(&self, config: Config) -> Result<()>;
 }
 
 /// The Data command is the one we invoke to get a set of metrics and geometry
@@ -89,16 +89,14 @@ impl DataCommand{
         } 
 
         metric_ids
-
-    
     }
 }
 
 impl RunCommand for DataCommand {
-    async fn run(&self) -> Result<()> {
+    async fn run(&self, config: Config) -> Result<()> {
         info!("Running `data` subcommand");
 
-        let popgetter = Popgetter::new().await?;
+        let popgetter = Popgetter::new_with_config(config).await?;
         let data_request = DataRequestSpec::from(self);
         let mut results = popgetter.get_data_request(&data_request).await?;
 
@@ -152,7 +150,7 @@ pub struct MetricsCommand {
 }
 
 impl RunCommand for MetricsCommand {
-    async fn run(&self) -> Result<()> {
+    async fn run(&self, config: Config) -> Result<()> {
         info!("Running `metrics` subcommand");
         Ok(())
     }
@@ -164,8 +162,8 @@ impl RunCommand for MetricsCommand {
 pub struct CountriesCommand;
 
 impl RunCommand for CountriesCommand {
-    async fn run(&self) -> Result<()> {
-        let _popgetter = Popgetter::new().await?;
+    async fn run(&self, config: Config) -> Result<()> {
+        let _popgetter = Popgetter::new_with_config(config).await?;
         Ok(())
     }
 }
@@ -176,7 +174,7 @@ impl RunCommand for CountriesCommand {
 pub struct SurveysCommand;
 
 impl RunCommand for SurveysCommand {
-    async fn run(&self) -> Result<()> {
+    async fn run(&self, config: Config) -> Result<()> {
         info!("Running `surveys` subcommand");
         Ok(())
     }
